generated-code/generated-v3/backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
python-jose[cryptography]
passlib[bcrypt]
dotenv